+++ 12/12-2.c
@@ -152,7 +152,7 @@ int main(int argc, char *argv[]) {
 void read_process_info(pid_t pid, Process_info **processes, unsigned long *num_processes) {
     char status_filename[STATUS_FILENAME_LENGTH];
     int snprintf_result = snprintf(
-        status_filename, sizeof(status_filename), "/proc/%d/status", pid);
+        status_filename, sizeof(status_filename), "/proc/%ld/status", (long)pid);
     if (snprintf_result < 0) {
         errExit("snprintf");
     } else if (snprintf_result >= sizeof(status_filename)) {
